# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Suvana Constructions API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactFormBody {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  phone: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  honeypot?: string;
}

export type QuoteFormBodyProjectType = typeof QuoteFormBodyProjectType[keyof typeof QuoteFormBodyProjectType];


export const QuoteFormBodyProjectType = {
  'kitchen-remodeling': 'kitchen-remodeling',
  drywall: 'drywall',
  plumbing: 'plumbing',
  electrical: 'electrical',
  flooring: 'flooring',
  fireplace: 'fireplace',
  basement: 'basement',
  painting: 'painting',
  remodeling: 'remodeling',
  handyman: 'handyman',
} as const;

export type QuoteFormBodyBudget = typeof QuoteFormBodyBudget[keyof typeof QuoteFormBodyBudget];


export const QuoteFormBodyBudget = {
  'under-5k': 'under-5k',
  '5k-15k': '5k-15k',
  '15k-50k': '15k-50k',
  '50k-100k': '50k-100k',
  'over-100k': 'over-100k',
  'not-sure': 'not-sure',
} as const;

export type QuoteFormBodyTimeline = typeof QuoteFormBodyTimeline[keyof typeof QuoteFormBodyTimeline];


export const QuoteFormBodyTimeline = {
  asap: 'asap',
  '1-3-months': '1-3-months',
  '3-6-months': '3-6-months',
  '6-12-months': '6-12-months',
  flexible: 'flexible',
} as const;

export interface QuoteFormBody {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  phone: string;
  projectType: QuoteFormBodyProjectType;
  /**
     * @minLength 2
     * @maxLength 200
     */
  location: string;
  budget?: QuoteFormBodyBudget;
  timeline?: QuoteFormBodyTimeline;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  honeypot?: string;
  recaptchaToken: string;
}

export interface SubmissionResult {
  success: boolean;
  id: number;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type SubmissionType = typeof SubmissionType[keyof typeof SubmissionType];


export const SubmissionType = {
  contact: 'contact',
  quote: 'quote',
} as const;

export interface Submission {
  id: number;
  type: SubmissionType;
  name: string;
  email: string;
  phone: string;
  projectType?: string | null;
  location?: string | null;
  budget?: string | null;
  timeline?: string | null;
  message: string;
  createdAt: string;
}

export interface SubmissionsPage {
  submissions: Submission[];
  total: number;
  page: number;
  limit: number;
}

export type SubmissionStatsByProjectTypeItem = {
  projectType: string;
  count: number;
};

export interface SubmissionStats {
  totalContacts: number;
  totalQuotes: number;
  recentSubmissions: number;
  byProjectType?: SubmissionStatsByProjectTypeItem[];
}

export interface Comment {
  id: number;
  submissionId: number;
  content: string;
  isShared: boolean;
  createdAt: string;
}

export interface AddCommentBody {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  content: string;
  isShared?: boolean;
}

export type GetSubmissionsParams = {
type?: GetSubmissionsType;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type GetSubmissionsType = typeof GetSubmissionsType[keyof typeof GetSubmissionsType];


export const GetSubmissionsType = {
  contact: 'contact',
  quote: 'quote',
  all: 'all',
} as const;