# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Suvana Constructions API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions/{id}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: Get all comments for a submission
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "404":
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment to a submission
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentBody"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /submissions/{id}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: commentId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Comment deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Comment:
      type: object
      required: [id, submissionId, content, isShared, createdAt]
      properties:
        id:
          type: integer
        submissionId:
          type: integer
        content:
          type: string
        isShared:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AddCommentBody:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 5000
        isShared:
          type: boolean
          default: false