# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Suvana Constructions API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Submit a contact inquiry
 */
export const submitContactBodyNameMin = 2;
export const submitContactBodyNameMax = 100;

export const submitContactBodyPhoneMin = 7;
export const submitContactBodyPhoneMax = 20;

export const submitContactBodyMessageMin = 10;
export const submitContactBodyMessageMax = 2000;



export const SubmitContactBody = zod.object({
  "name": zod.string().min(submitContactBodyNameMin).max(submitContactBodyNameMax),
  "email": zod.string().email(),
  "phone": zod.string().min(submitContactBodyPhoneMin).max(submitContactBodyPhoneMax),
  "message": zod.string().min(submitContactBodyMessageMin).max(submitContactBodyMessageMax),
  "honeypot": zod.string().optional()
})


/**
 * @summary Submit a quote request
 */
export const submitQuoteBodyNameMin = 2;
export const submitQuoteBodyNameMax = 100;

export const submitQuoteBodyPhoneMin = 7;
export const submitQuoteBodyPhoneMax = 20;

export const submitQuoteBodyLocationMin = 2;
export const submitQuoteBodyLocationMax = 200;

export const submitQuoteBodyMessageMin = 10;
export const submitQuoteBodyMessageMax = 2000;



export const SubmitQuoteBody = zod.object({
  "name": zod.string().min(submitQuoteBodyNameMin).max(submitQuoteBodyNameMax),
  "email": zod.string().email(),
  "phone": zod.string().min(submitQuoteBodyPhoneMin).max(submitQuoteBodyPhoneMax),
  "projectType": zod.enum(['kitchen-remodeling', 'drywall', 'plumbing', 'electrical', 'flooring', 'fireplace', 'basement', 'painting', 'remodeling', 'handyman']),
  "location": zod.string().min(submitQuoteBodyLocationMin).max(submitQuoteBodyLocationMax),
  "budget": zod.enum(['under-5k', '5k-15k', '15k-50k', '50k-100k', 'over-100k', 'not-sure']).optional(),
  "timeline": zod.enum(['asap', '1-3-months', '3-6-months', '6-12-months', 'flexible']).optional(),
  "message": zod.string().min(submitQuoteBodyMessageMin).max(submitQuoteBodyMessageMax),
  "honeypot": zod.string().optional(),
  "recaptchaToken": zod.string()
})


/**
 * @summary Get all form submissions (admin)
 */
export const getSubmissionsQueryPageDefault = 1;

export const getSubmissionsQueryLimitDefault = 20;
export const getSubmissionsQueryLimitMax = 100;



export const GetSubmissionsQueryParams = zod.object({
  "type": zod.enum(['contact', 'quote', 'all']).optional(),
  "page": zod.coerce.number().min(1).default(getSubmissionsQueryPageDefault),
  "limit": zod.coerce.number().min(1).max(getSubmissionsQueryLimitMax).default(getSubmissionsQueryLimitDefault)
})

export const GetSubmissionsResponse = zod.object({
  "submissions": zod.array(zod.object({
  "id": zod.number(),
  "type": zod.enum(['contact', 'quote']),
  "name": zod.string(),
  "email": zod.string(),
  "phone": zod.string(),
  "projectType": zod.string().nullish(),
  "location": zod.string().nullish(),
  "budget": zod.string().nullish(),
  "timeline": zod.string().nullish(),
  "message": zod.string(),
  "createdAt": zod.coerce.date()
})),
  "total": zod.number(),
  "page": zod.number(),
  "limit": zod.number()
})


/**
 * @summary Get submission statistics
 */
export const GetSubmissionStatsResponse = zod.object({
  "totalContacts": zod.number(),
  "totalQuotes": zod.number(),
  "recentSubmissions": zod.number(),
  "byProjectType": zod.array(zod.object({
  "projectType": zod.string(),
  "count": zod.number()
})).optional()
})


/**
 * @summary Get all comments for a submission
 */
export const GetCommentsParams = zod.object({
  "id": zod.coerce.number()
})

export const GetCommentsResponseItem = zod.object({
  "id": zod.number(),
  "submissionId": zod.number(),
  "content": zod.string(),
  "isShared": zod.boolean(),
  "createdAt": zod.coerce.date()
})
export const GetCommentsResponse = zod.array(GetCommentsResponseItem)


/**
 * @summary Add a comment to a submission
 */
export const AddCommentParams = zod.object({
  "id": zod.coerce.number()
})

export const addCommentBodyContentMax = 5000;

export const addCommentBodyIsSharedDefault = false;

export const AddCommentBody = zod.object({
  "content": zod.string().min(1).max(addCommentBodyContentMax),
  "isShared": zod.boolean().default(addCommentBodyIsSharedDefault)
})


/**
 * @summary Delete a comment
 */
export const DeleteCommentParams = zod.object({
  "id": zod.coerce.number(),
  "commentId": zod.coerce.number()
})